Adding a NOTICE file to cover non-CLA commits.
I was not able to reach one of the commit authors for signing the CNCF
CLA, so adding a NOTICE file that references section 7 of the CNCF CLA
instead. (This is not a donated repo, but the code is being forked from
another repository for further development, which I think should be
covered by section 7 as well).