feat(index): restructure layout into a grid
Mimick the layout of the old UI by rendering a sorted grid of dashboard
groups and ungrouped dashboards. On-hover of a dashboard group, a list
of dashboards is shown. Clicking on a dashboard group should navigate to
a new component: dashboard-group-summary. Clicking on an ungrouped
dashboard navigates to the dashboard-summary component.

Signed-off-by: Brady Pratt <bpratt@redhat.com>