fix(navigation): support both group and dashboard endpoints
add a top level summary component that determines whether to render the
dashboard group summary or to render the dashboard summary. This seems a
bit inefficient but this is the behavior of the old UI. The index sets
the type based on the griditem then we confirm when loading the summary.
We still need to do this check in case the endpoint is navigated to
directly

Signed-off-by: Brady Pratt <bpratt@redhat.com>